pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alberta-framework"
version = "0.8.0"
description = "Implementation of the Alberta Plan for AI Research - continual learning with meta-learned step-sizes"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.13"
authors = [
    { name = "J. Klawson", email = "j-klawson@users.noreply.github.com" }
]
keywords = ["machine-learning", "continual-learning", "meta-learning", "jax", "reinforcement-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "jax>=0.4",
    "jaxlib>=0.4",
    "numpy>=2.0",
    "matplotlib>=3.8",
    "scipy>=1.11",
    "joblib>=1.3",
    "tqdm>=4.66",
    "chex>=0.1.86",
    "jaxtyping>=0.2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
    "mypy",
    "beartype>=0.18.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
    "mkdocs-gen-files>=0.5",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-section-index>=0.3",
]
gymnasium = [
    "gymnasium>=0.29.0",
]
bsuite = [
    "dm-env>=1.6",
    "optax>=0.1.7",
    "dm-haiku>=0.0.10",
    "plotnine>=0.12",
]
gpu = [
    "jax[cuda12]>=0.4",
]

[project.urls]
Homepage = "https://github.com/j-klawson/alberta-framework"
Repository = "https://github.com/j-klawson/alberta-framework"
Documentation = "https://j-klawson.github.io/alberta-framework"
Issues = "https://github.com/j-klawson/alberta-framework/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/alberta_framework"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
# F722: Syntax error in forward annotation - ruff doesn't understand jaxtyping shape annotations
ignore = ["F722"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_ignores = true
strict = true

[[tool.mypy.overrides]]
module = "gymnasium.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "matplotlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "scipy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "joblib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tqdm.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chex.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jaxtyping.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dm_env.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bsuite.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "haiku.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "optax.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "plotnine.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "termcolor.*"
ignore_missing_imports = true

# Benchmarks use dynamic dm_env/bsuite/haiku types that mypy can't fully resolve
[[tool.mypy.overrides]]
module = "benchmarks.*"
disable_error_code = ["call-arg", "arg-type", "union-attr", "attr-defined", "operator"]

# chex dataclasses don't have full mypy support - disable strict checks for core modules
[[tool.mypy.overrides]]
module = "alberta_framework.core.types"
disable_error_code = ["call-arg"]

[[tool.mypy.overrides]]
module = "alberta_framework.core.normalizers"
disable_error_code = ["call-arg"]

[[tool.mypy.overrides]]
module = "alberta_framework.core.optimizers"
disable_error_code = ["call-arg"]

[[tool.mypy.overrides]]
module = "alberta_framework.core.learners"
disable_error_code = ["call-arg", "arg-type", "union-attr"]

[[tool.mypy.overrides]]
module = "alberta_framework.core.multi_head_learner"
disable_error_code = ["call-arg", "arg-type", "union-attr"]

[[tool.mypy.overrides]]
module = "alberta_framework.streams.synthetic"
disable_error_code = ["call-arg"]

[[tool.mypy.overrides]]
module = "alberta_framework.streams.gymnasium"
disable_error_code = ["call-arg"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
